Extracted backup functionality
