+++ app/src/main/java/com/futsch1/medtimer/OptionsMenu.java
@@ -16,13 +16,9 @@ import android.widget.Toast;
 import androidx.activity.result.ActivityResultLauncher;
 import androidx.annotation.NonNull;
 import androidx.appcompat.view.menu.MenuBuilder;
-import androidx.core.content.FileProvider;
 import androidx.core.view.MenuProvider;
 import androidx.navigation.NavController;
 
-import com.futsch1.medtimer.database.JSONBackup;
-import com.futsch1.medtimer.database.JSONMedicineBackup;
-import com.futsch1.medtimer.database.JSONReminderEventBackup;
 import com.futsch1.medtimer.exporters.CSVExport;
 import com.futsch1.medtimer.exporters.Exporter;
 import com.futsch1.medtimer.exporters.PDFExport;
@@ -33,18 +29,16 @@ import com.futsch1.medtimer.reminders.ReminderProcessor;
 import java.io.File;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.net.URLConnection;
-import java.util.List;
 import java.util.TimeZone;
 
 public class OptionsMenu implements MenuProvider {
     private final Context context;
     private final MedicineViewModel medicineViewModel;
-    private final HandlerThread backgroundThread;
     private final ActivityResultLauncher<Intent> openFileLauncher;
     private final NavController navController;
-    private PendingFileOperation pendingFileOperation = PendingFileOperation.NONE;
+    private final HandlerThread backgroundThread;
     private Menu menu;
+    private BackupManager backupManager;
 
     public OptionsMenu(Context context, MedicineViewModel medicineViewModel, ActivityResultLauncher<Intent> openFileLauncher, NavController navController) {
         this.context = context;
@@ -53,6 +47,27 @@ public class OptionsMenu implements MenuProvider {
         this.navController = navController;
         backgroundThread = new HandlerThread("Export");
         backgroundThread.start();
+    }
+
+    public void onDestroy() {
+        backupManager.onDestroy();
+        backgroundThread.quitSafely();
+    }
+
+    @Override
+    public void onCreateMenu(@NonNull Menu menu, @NonNull MenuInflater menuInflater) {
+        menuInflater.inflate(R.menu.main, menu);
+        menu.setGroupDividerEnabled(true);
+        enableOptionalIcons(menu);
+
+        this.backupManager = new BackupManager(context, menu, medicineViewModel, openFileLauncher);
+        this.menu = menu;
+        setupSettings();
+        setupVersion();
+        setupAppURL();
+        setupClearEvents();
+        setupExport();
+        setupGenerateTestData();
     }
 
     @SuppressLint("RestrictedApi")
@@ -66,93 +81,6 @@ public class OptionsMenu implements MenuProvider {
                 Log.e("Menu", "onMenuOpened", e);
             }
         }
-    }
-
-    public void onDestroy() {
-        backgroundThread.quitSafely();
-    }
-
-    public void fileSelected(Uri data) {
-        String json = FileHelper.readFromUri(data, context.getContentResolver());
-        boolean restoreSuccessful = false;
-        if (json != null) {
-            if (pendingFileOperation == PendingFileOperation.MEDICINE) {
-                restoreSuccessful = restoreBackup(json, new JSONMedicineBackup());
-            } else if (pendingFileOperation == PendingFileOperation.REMINDER_EVENTS) {
-                restoreSuccessful = restoreBackup(json, new JSONReminderEventBackup());
-            }
-
-            pendingFileOperation = PendingFileOperation.NONE;
-        }
-
-        new AlertDialog.Builder(context)
-                .setMessage(restoreSuccessful ? R.string.restore_successful : R.string.restore_failed)
-                .setPositiveButton(R.string.ok, (dialog, which) -> {
-                    // Intentionally empty
-                })
-                .show();
-    }
-
-    private <T> boolean restoreBackup(String json, JSONBackup<T> backup) {
-        List<T> backupData = backup.parseBackup(json);
-        if (backupData != null) {
-            backup.applyBackup(backupData, medicineViewModel.medicineRepository);
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public void onCreateMenu(@NonNull Menu menu, @NonNull MenuInflater menuInflater) {
-        menuInflater.inflate(R.menu.main, menu);
-        menu.setGroupDividerEnabled(true);
-        enableOptionalIcons(menu);
-
-        this.menu = menu;
-        setupSettings();
-        setupVersion();
-        setupAppURL();
-        setupClearEvents();
-        setupExport();
-        setupGenerateTestData();
-        setupBackup();
-        setupEventsBackup();
-    }
-
-    @Override
-    public boolean onMenuItemSelected(@NonNull MenuItem menuItem) {
-        return false;
-    }
-
-    private void setupEventsBackup() {
-        Handler handler = new Handler(backgroundThread.getLooper());
-
-        MenuItem item = menu.findItem(R.id.backup_events);
-        item.setOnMenuItemClickListener(menuItem -> {
-            handler.post(() ->
-                    createBackup(new JSONReminderEventBackup(),
-                            medicineViewModel.medicineRepository.getAllReminderEvents(),
-                            "ReminderEvents"));
-            return true;
-        });
-
-        item = menu.findItem(R.id.restore_events);
-        item.setOnMenuItemClickListener(menuItem -> {
-            new AlertDialog.Builder(context)
-                    .setTitle(R.string.restore)
-                    .setMessage(R.string.restore_events_start)
-                    .setCancelable(true)
-                    .setPositiveButton(R.string.ok, (dialog, which) -> openBackup(PendingFileOperation.REMINDER_EVENTS))
-                    .setNegativeButton(R.string.cancel, (dialog, which) -> { // Intentionally left empty
-                    }).show();
-            return true;
-        });
-    }
-
-    private <T> void createBackup(JSONBackup<T> jsonBackup, List<T> backupData, String backupType) {
-        String jsonContent = jsonBackup.createBackup(medicineViewModel.medicineRepository.getVersion(),
-                backupData);
-        createAndSave(jsonContent, backupType);
     }
 
     private void setupSettings() {
@@ -226,70 +154,22 @@ public class OptionsMenu implements MenuProvider {
         }
     }
 
-    private void setupBackup() {
-        Handler handler = new Handler(backgroundThread.getLooper());
-
-        MenuItem item = menu.findItem(R.id.medicine_backup);
-        item.setOnMenuItemClickListener(menuItem -> {
-            handler.post(() ->
-                    createBackup(new JSONMedicineBackup(),
-                            medicineViewModel.medicineRepository.getMedicines(),
-                            "Medicine"));
-            return true;
-        });
-
-        item = menu.findItem(R.id.restore_medicine_backup);
-        item.setOnMenuItemClickListener(menuItem -> {
-            new AlertDialog.Builder(context)
-                    .setTitle(R.string.restore)
-                    .setMessage(R.string.restore_start)
-                    .setCancelable(true)
-                    .setPositiveButton(R.string.ok, (dialog, which) -> openBackup(PendingFileOperation.MEDICINE))
-                    .setNegativeButton(R.string.cancel, (dialog, which) -> { // Intentionally left empty
-                    }).show();
-            return true;
-        });
-    }
-
     private void export(Exporter exporter) {
         File csvFile = new File(context.getCacheDir(), PathHelper.getExportFilename(exporter));
         try {
             exporter.export(csvFile);
-            shareFile(csvFile);
+            FileHelper.shareFile(context, csvFile);
         } catch (Exporter.ExporterException e) {
             Toast.makeText(context, R.string.export_failed, Toast.LENGTH_LONG).show();
         }
     }
 
-    private void createAndSave(String fileContent, String backupType) {
-        File file = new File(context.getCacheDir(), PathHelper.getBackupFilename(backupType));
-        if (FileHelper.saveToFile(file, fileContent)) {
-            shareFile(file);
-        } else {
-            Toast.makeText(context, R.string.backup_failed, Toast.LENGTH_LONG).show();
-        }
+    @Override
+    public boolean onMenuItemSelected(@NonNull MenuItem menuItem) {
+        return false;
     }
 
-    private void openBackup(PendingFileOperation pendingFileOperation) {
-        this.pendingFileOperation = pendingFileOperation;
-        Intent intent = new Intent(Intent.ACTION_OPEN_DOCUMENT);
-        intent.addCategory(Intent.CATEGORY_OPENABLE);
-        intent.setType("application/json");
-        openFileLauncher.launch(intent);
+    public void fileSelected(Uri data) {
+        backupManager.fileSelected(data);
     }
-
-    private void shareFile(File file) {
-        Uri uri = FileProvider.getUriForFile(context, "com.futsch1.medtimer.fileprovider", file);
-        Intent intentShareFile = new Intent(Intent.ACTION_SEND);
-
-        intentShareFile.setDataAndType(uri, URLConnection.guessContentTypeFromName(file.getName()));
-        //Allow sharing apps to read the file Uri
-        intentShareFile.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-        //Pass the file Uri instead of the path
-        intentShareFile.putExtra(Intent.EXTRA_STREAM,
-                uri);
-        context.startActivity(Intent.createChooser(intentShareFile, "Share File"));
-    }
-
-    private enum PendingFileOperation {NONE, MEDICINE, REMINDER_EVENTS}
 }
